fix(deps): update uppy monorepo

+++ packages/web-pkg/src/services/uppy/uppyService.ts
@@ -276,6 +276,9 @@ export class UppyService {
       this.clearInputs()
     })
     this.uppy.on('complete', (result) => {
+      if (!result) {
+        return
+      }
       this.publish('uploadCompleted', result)
       result.successful.forEach((file: any) => {
         this.uppy.removeFile(file.id)
